make memory layout conversions explicit
Summary:

The UX of doing the layout conversions out the hood is nice, but
removing the magic handling for now and spelling these out explicitly
to make it clearer from the code where expensive things are happening.

IMO this will make it easier to reason about how to speed up our code
as we won't have to rely on traces to understand where the contiguous
calls are.

Test Plan:

```
with-proxy ./tests/test_everything.sh
```

Reviewers:

Subscribers:

Tasks:

Tags: