enable autocast + compile + FSDP + Float8Linear (#172)
Summary:
This adds a couple of config options to unbreak autocast + compile + FSDP + Float8Linear. To enable these options, the user needs to do:

```
config.enable_amax_init = False
config.enable_pre_and_post_forward = False
```

The `enable_amax_init` config adds the option to disable amax initialization. The reason this is currently broken is:
1. FSDP is not full-graph friendly (regardless of compile)
2. the amax init function has a graph break in distributed code because it uses inplace distributed collectives.  I did try to use functional collectives (https://github.com/pytorch-labs/float8_experimental/pull/171), but that ran into numerical issues with compile, so for now just working around it.
3. graph breaks in Float8Linear code are not supported because of the issue documented in https://github.com/pytorch-labs/float8_experimental/pull/166
4. so, as a workaround for all of the above, we just skip amax init for now.  We do know from NVIDIA that this path is not needed for model convergence, and TE does not support this at all. It was nice for testing but not necessary for training jobs.

The second config option disables pre-forward and post-forward. I don't have a repro in a unit test for now, but this does unbreak LLaMa 7B on 8 GPUs with FSDP + compile. Specifically, the thing which is broken in pre-forward/post-forward is assignment on module attributes. My hunch is that this graph breaks if autocast + FSDP are on, and graph breaks are not supported due to (3) above.

Pull Request resolved: https://github.com/pytorch-labs/float8_experimental/pull/172

Test Plan:
```
// unit / integration tests
with-proxy test/test_everything.sh

// run the LLaMa 7b trainer on 8 GPUs with autocast + compile + FSDP + Float8Linear, no compile errors
```

Reviewed By: drisspg

Differential Revision: D52468625

Pulled By: vkuzo

fbshipit-source-id: be4fac927b8520602ed018e96d7a49056e9c6e06