Add tests for scaled_mm + DTensor (#209)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/float8_experimental/pull/209

This PR adds test to test the scaled_mm op working with DTensor.

Currently using the same style of test runner
compared to other tests in float8_experimental. We shall consider
switching to DTensorTestBase or MultiProcessTestCase in the repo but
didn't do it in this PR

imported-using-ghimport

Test Plan: Imported from OSS

Reviewed By: drisspg

Differential Revision: D53731879

Pulled By: wanchaol

fbshipit-source-id: a1e8fad87166b67148bc32304da68d2497b9721d