early return by check tensor already casted or not (#233)
Summary:
as titled, it turns out we don't need to install additional hooks base on our TP + FP8 design. The only thing we need to do here is to be able to turn off activation casting, so that we can put activation casting in the TP hooks

So just check if the tensor already been casted to fp8 or not, as in TP we would cast activation into the DTensor's Float8Colwise/Rowwise instead.

Pull Request resolved: https://github.com/pytorch-labs/float8_experimental/pull/233

Reviewed By: drisspg

Differential Revision: D54497744

Pulled By: wanchaol

fbshipit-source-id: 20c4f6799bf91716778e2257388cb53b13373064