delayed scaling: stop syncing weight amax values across ranks
Summary:

FSDP already ensures that each rank receives the same weight, so the
amaxes of weights are the same on each rank.

I checked performance before/after on the multi GPU benchmark and
didn't see a significant impact on the toy model, but less comms value is better.

Test Plan:

./test_everything.sh passes

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 615a026367606ef7e5c9922e4c6d5bd1e729ad0e
Pull Request resolved: https://github.com/pytorch-labs/float8_experimental/pull/277