delayed scaling: delete Float8LinearMixin (#276)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/float8_experimental/pull/276

The mixin was originally used to share code with Float8 versions of
RowParallelLinear and ColParallelLinear.  Since we moved those to
DTensor, the mixin is not needed anymore.  Removing it to
simplify the code in preparation of upcoming delayed scaling
improvements.

In addition, making the from_float conversion use meta device
to speed it up.

Reviewed By: drisspg

Differential Revision: D58396926

fbshipit-source-id: 353fbd8c5c532f670391f1a1620cbc2bd034b42b