add per-gemm config to `Float8LinearConfig`
Summary:

Previously the per-gemm configuration had to be hardcoded in library
code. This PR exposes it to the top-level UX by adding a
`Float8GemmConfig` field to `Float8LinearConfig`.

Note that today the only supported configuration option is
`use_fast_accum`.  In the future, configuring output_dtype
and whether to keep a gemm in higher precision would go here.

Test Plan:

```
./test/test_everything.sh
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: f6439de810c42612190d3be07b93e46528b0228f
Pull Request resolved: https://github.com/pytorch-labs/float8_experimental/pull/334