bring back torch.autograd.Function for float8 matmul
Summary:

This is a redo of
https://github.com/pytorch-labs/float8_experimental/pull/316

With upcoming support of scaling granularities other than tensorwise,
we need a good way to control which gemm kernel to call and how to scale
the input tensors in fwd and bwd. A `torch.autograd.Function` override
is the cleanest way to do that, and in 2024 this now works with
`torch.compile`.

Test Plan:

```
./test/test_everything.sh
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 3992ccc9fcb764ecc3641e4b32cccef93233e9f4
Pull Request resolved: https://github.com/pytorch-labs/float8_experimental/pull/336