add per-gemm config to `Float8LinearConfig` (#334)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/float8_experimental/pull/334

Previously the per-gemm configuration had to be hardcoded in library
code. This PR exposes it to the top-level UX by adding a
`Float8GemmConfig` field to `Float8LinearConfig`.

Note that today the only supported configuration option is
`use_fast_accum`.  In the future, configuring output_dtype
and whether to keep a gemm in higher precision would go here.

Reviewed By: weifengpy

Differential Revision: D60252069

fbshipit-source-id: bca34eb49e1bf046f937e32b11b2369b535d56e6