bring back torch.autograd.Function for float8 matmul (#344)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/float8_experimental/pull/344

This is a redo of
https://github.com/pytorch-labs/float8_experimental/pull/316

With upcoming support of scaling granularities other than tensorwise,
we need a good way to control which gemm kernel to call and how to scale
the input tensors in fwd and bwd. A `torch.autograd.Function` override
is the cleanest way to do that, and in 2024 this now works with
`torch.compile`.

Reviewed By: drisspg

Differential Revision: D60291446

fbshipit-source-id: 472f392227bca1c7f83ea0c1234285bc576e58d2