clean up casting functions: delayed scaling (#346)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/float8_experimental/pull/346

Removes delayed scaling from `float8_tensor.py`. After this PR, the
invariant is that everything in `float8_tensor.py` requires the scale to
be calculated elsewhere. This moves the codebase towards separation of
concerns for calculating the scale (via various scaling strategies),
separated from creating an instance of `Float8Tensor`.

Note that stateful delayed scaling is the reason we need this separation.

Reviewed By: drisspg

Differential Revision: D60291447

fbshipit-source-id: 79644cc7857311fff37f4c0238aba78e20ad5ac3